docs/filesystem: Weaken stateoverlay wording and direct to alternatives
We wrote the code and it is useful. We are probably just going
to need to support it into the forseeable future. However, my
core problem with it is what I wrote in the docs here: it has
no equivalent in the docker/podman ecosystem, which I think
cuts against a core principle that we're trying to make
the booted host align with containers.

Of course, stateoverlay isn't a lot of code, and in theory
it's a bit independent of ostree, so perhaps we could try to
split it out at some point and then it could be used by e.g.
`podman run --mount=type=stateoverlay` or so.

Before anyone starts just blindly enabling stateoverlays
I'd like them to have considered alternatives such as the
symlink approach.

Signed-off-by: Colin Walters <walters@verbum.org>