store: Don't use CONSUME flag
libostree currently has a bug here with fsverity support;
in the consume case (which right now is always zero-sized
files because we don't hardlink them) it doesn't enable verity.

This would be an easy bug to fix in libostree.

But OTOH there's no real reason to set consume here either.
The main idea of consume is when one is operating on potentially
large files external to libostree, but most things here are
hardlinked, and what isn't is just the zero sized files.

Signed-off-by: Colin Walters <walters@verbum.org>