lsm, imgstorage: Rework relabeling
The previous work here wasn't quite right in a few ways.
Our LSM/SELinux code is a bit complex and under-tested.

Here we:

First, refactor some of the labeling bits so we have a clean
"relabel this file" API.

For the bootc-owned containers-storage we don't want
"recursive create dir and relabel" in the general case - we
need to handle upgrades, where there are definitely
non-directories too.

Hence rework the API to just be a clean recursive
relabeling pass, don't attempt to create anything
on our own.

The install path hence changes to let podman create
the dirs first, then we relabel.

While we're here:

- Rework the recursive traversal to operate on shared
  single `&mut` path buffers to avoid a heap alloc per directory.
- Add a `bootc internals relabel` CLI verb that
  makes it easy to test this code both interactively
  and in integration testing.
- Add a test case

Closes: https://github.com/bootc-dev/bootc/issues/1219

Signed-off-by: Colin Walters <walters@verbum.org>