imgstore: Fallback to container root sepolicy
Prior to this, get_ensure_imgstore would fail when run from a non-bootc
system. Instead, in an attempt to make this function idempotent, let's
fallback to the container root sepolicy instead of failing. This should
only happen when running cleanup() during a to-existing install (i.e.
there is not yet a bootc system).

Signed-off-by: ckyrouac <ckyrouac@redhat.com>