store: Improve logging for filtered content
We thought this code may be related to a bug where layer
content was apparently getting lost, though
I don't think it was in the end.

Anyways though, use the same "show first n items from iterator"
code that we have in the lint to actually display the filtered
state.

Signed-off-by: Colin Walters <walters@verbum.org>