install: Do a dynamic mount for /var/tmp
Closes: https://github.com/bootc-dev/bootc/issues/1292

Basically we were doing the `/proc/1/root/var/tmp` trick for
`/var/tmp` because we didn't have the dynamic bind mount infrastructure
before. Now we do, so use it instead.

The specific motivation is that Go in some cases uses `EvalSymlinks`
which gets confused by the `/proc/<pid>/root` magic links.

Also, this deletes a lot of code.

Signed-off-by: Colin Walters <walters@verbum.org>