spec: Note we don't canonicalize oci
The previous code in trying to parse `oci` was wrong; the syntax
for an `oci` transport is the same as oci-archive, which
includes a file path (and there's no mechanism to quote `:` note).

The canonical logic for all of this stuff is in Go, there's
no canonical Rust library (yet, though I did think about putting
this in oci-spec).

Previous to this attempt to handle tagged+digested, we weren't
parsing image references at all.

First, factor out a `canonicalize_reference` helper since that's
what we're really doing here, it's independent of the *transport*.

Fix the canonicalize function to drop out trying to parse `oci`.

Add a separate test case that incorrectly passes just so it's
a bit more obvious to fix this later.

Note that today at least `skopeo` rejects trying to fetch
via tagged+digested form from an `oci:` so it's fine if
we don't canonicalize here yet, even though it could confuse
someone.

Signed-off-by: Colin Walters <walters@verbum.org>