Dedup sepolicy handling
For historical reasons the ostree sepolicy API can exist as a
no-op even if it didn't find a policy, one has to query `.csum()`
or `.name()` to verify it's present.

In our code just map that case to None.

Followup to 99d30dfd556866091c558559810647bdde4e1ee1
to ensure we consistently handle this case.

Signed-off-by: Colin Walters <walters@verbum.org>