lints: Add --no-truncate flag to bootc container lint
Useful when running automatic reports and analyses
where all issues need to be visible at once without truncation.

To do this we needed to add a LintExecutionConfig to
propagate configuration through the lint system.

Also, refactor how we print things so there's a
shared helper functions for consistent formatting with and without truncation.

Closes: https://github.com/bootc-dev/bootc/issues/1260

Signed-off-by: Colin Walters <walters@verbum.org>