Create /run/reboot-required by default
When staging a new deployment, create /run/reboot-required to signal that
a reboot is needed. This file is monitored by kured (Kubernetes Reboot
Daemon) and other tools to detect when a system needs to be rebooted.

This makes it easier to integrate bootc with kured and similar tools
without requiring manual configuration or bridging.

Signed-off-by: Colin Walters <walters@verbum.org>