lib/src/reboot: use systemd-run to execute the reboot
When we do a reboot it is triggered inside the bootc namespace.
As we implement support for soft-reboots we need to make sure
that systemd has a view into the mounted /run/nextroot
to be able to act on doing a soft-reboot or a reboot.
By using systemd-run we avoid the limited view in the current
namespace.

Signed-off-by: Joseph Marrero Corchado <jmarrero@redhat.com>