ostree-ext/store: use BTreeMap for merge commit metadata
`HashMap` is unordered which means that the conversion to GVariant
(and thus its serialization) is not reproducible since the conversion
basically just maps to an GVariant array. That in turns means the commit
hash itself is not reproducible.

All we really need to do is just swap it for a `BTreeMap` since all the
types involved here implement ordering.