ostree-ext/store: don't include filtered files in metadata
Filtered files are only determined at the time we import a layer.
So if that layer is already imported, we won't have that information
available. That in turn means that the metadata is state-dependent,
which in turn means that the commit digest is not reproducible.

We still want to provide the filtered files warning though. Just make
this information part of the LayeredImageState object instead. The
obvious downside of that is that now we only get that warning the first
time the layer is imported and it's no longer part of the commit object
itself.

One way to make this more sticky is to attach it to the individual
layers' commits instead, and then the merge commit can coalesce them.

Related: https://github.com/bootc-dev/bootc/issues/1346