ostree-ext/cli: add `--digestfile` to `ostree container image pull`
This is something that I've wanted many times. There's no easy way from
e.g. bash to convert an image ref into the escaped OSTree ref and so
it's always awkward to be able to tell what ref/commit was pulled down.

Let's just match podman here and add a `--digestfile` switch for this.