tree-wide: Clean up our re-exec path
Prep for fixing https://github.com/bootc-dev/bootc/issues/1434

Basically in the selinux path we copy our binary to a tempfile,
which breaks `/proc/self/exe`.

Fix this by setting an environment variable when we do that re-exec
and ensuring that *everything* references an internal API (now
moved to utils/ so it can be shared) that looks for the env
var first.