kernel: Add value helpers
Many cases want to find the value of an argument, no need
to return the key back to them.

Also IMO `value_lossy()` is a HUGE trap because aside from
logging it's almost never the right thing to do. I plan
to remove it in a followup.

So this also adds a `utf8` helper that returns an error,
and changes one case to use it.

Signed-off-by: Colin Walters <walters@verbum.org>